+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   Features:
   * Support for MongoDB (using Mongoid).
   * config.feedback_poll is now deprecated, use config.apns.feedback_receiver.frequency instead.
+  * Add config.apns.feedback_receiver.enabled to optionally enabled the APNs feedback receiver (#129).
 
   Bug fixes:
   * Fix setting the log level when using Rails 4+ or without Rails (#124).